everything/BiosafetyEL3/requirements.txt
# Perfect for Apple Silicon M1/M2/M3/M4 – native MPS acceleration
torch==2.4.1
biopython>=1.83
numpy>=1.24
tqdm>=4.66
streamlit